+++ yorm/__init__.py
@@ -3,7 +3,7 @@
 import sys
 
 __project__ = 'YORM'
-__version__ = '0.4.dev16'
+__version__ = '0.4.dev17'
 
 VERSION = __project__ + '-' + __version__